youtube-italian-vercel/requirements.txt
youtube-transcript-api==0.6.1
deep-translator==1.11.4
gtts==2.5.0
python-docx==1.1.0
fpdf2==2.7.6
requests==2.31.0